# API/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow API Docs",
    "description": "Documentacion interactiva completa de la API de CoffeeFlow: Autenticacion, Usuarios, Mesas, Productos, Pedidos, Cocina (ingredientes/recetas/inventario/alertas), Caja (turnos/tickets/pagos/cortes/gastos/compras) y Notificaciones.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Servidor Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuarios",
        "parameters": [
          {
            "name": "rol",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "mesero",
                "cocinero",
                "cajero"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de usuarios"
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear usuario",
        "description": "Solo administrador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Roberto"
                  },
                  "apellido_paterno": {
                    "type": "string",
                    "example": "Gomez"
                  },
                  "apellido_materno": {
                    "type": "string",
                    "example": "Diaz"
                  },
                  "telefono": {
                    "type": "string",
                    "example": "4421234567"
                  },
                  "correo": {
                    "type": "string",
                    "example": "roberto.mesero@cafeteria.com"
                  },
                  "usuario": {
                    "type": "string",
                    "example": "mesero3"
                  },
                  "contrasena": {
                    "type": "string",
                    "example": "mesero123"
                  },
                  "rol": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "mesero",
                      "cocinero",
                      "cajero"
                    ],
                    "example": "mesero"
                  }
                },
                "required": [
                  "nombre",
                  "correo",
                  "usuario",
                  "contrasena",
                  "rol"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          },
          "409": {
            "description": "Correo o usuario ya en uso"
          }
        }
      }
    },
    "/api/usuarios/{id_usuario}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener usuario",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar usuario",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "telefono": {
                    "type": "string",
                    "example": "4429876543"
                  },
                  "correo": {
                    "type": "string",
                    "example": "nuevo.correo@cafeteria.com"
                  },
                  "activo": {
                    "type": "boolean",
                    "example": true
                  },
                  "contrasena": {
                    "type": "string",
                    "example": "nuevaClave123"
                  },
                  "rol": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "mesero",
                      "cocinero",
                      "cajero"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar usuario",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminado"
          },
          "400": {
            "description": "Un admin no puede eliminarse a si mismo"
          }
        }
      }
    },
...
  }
}